+++ source/Messaging.Application/Transactions/MoveIn/Notifications/CreateEndOfSupplyNotificationHandler.cs
@@ -16,10 +16,9 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using Messaging.Application.Transactions.MoveIn.Notifications;
 using Messaging.Domain.Transactions.MoveIn;
 
-namespace Messaging.Application.Transactions.MoveIn;
+namespace Messaging.Application.Transactions.MoveIn.Notifications;
 
 public class CreateEndOfSupplyNotificationHandler : IRequestHandler<CreateEndOfSupplyNotification, Unit>
 {
